Try to let all/search/filterPaginationView extends paginationView, but not working
